:hammer: [redux] 调整redux通用类型的声明
