+++ src/app/src/utils/ui-state-handler.ts
@@ -8,7 +8,7 @@ import TAlert from '../components/TApi/TAlert';
 import { hideAlert, hideModal } from '@src/shared/redux/actions/ui';
 import { Toast, Portal } from '@ant-design/react-native';
 import TModal from '../components/TApi/TModal';
-import { TRPGDispatch } from '@src/shared/redux/types/redux';
+import { TRPGDispatch } from '@src/shared/redux/types/__all__';
 
 type UIMap = Map<string, any>;
 interface FactoryOptions {
