+++ src/web/containers/main/group/GroupInfo.tsx
@@ -22,7 +22,7 @@ import {
   TRPGState,
   TRPGDispatch,
   TRPGDispatchProp,
-} from '@src/shared/redux/types/redux';
+} from '@src/shared/redux/types/__all__';
 import { AlertPayload } from '@src/shared/redux/types/ui';
 
 interface Props extends TRPGDispatchProp {
